FUSION/pyproject.toml
[project]
name = "fusion"
version = "1.0.0"
description = "FUSION - BEVAL & CONVEI Integration Layer"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "MAJECT Team"}
]
keywords = ["behavioral-analysis", "integration", "api", "metrics"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Web framework
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    
    # SocketIO client
    "python-socketio[asyncio-client]>=5.10.0",
    
    # HTTP client
    "httpx>=0.25.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    
    # Logging
    "colorlog>=6.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
fusion-api = "api.server:main"
fusion-collector = "integration.beval_collector:main_sync"
fusion-init-db = "db.init_db:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["fusion"]

[tool.black]
line-length = 100
target-version = ['py312']

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true

